debug sparselinear

+++ tensorgraph/layers/linear.py
@@ -105,8 +105,8 @@ class SparseLinear(Template):
     def _train_fprop(self, state_below):
         idx, val = state_below
         # nrow = val.get_shape()[0]
-        import pdb; pdb.set_trace()
-        X = tf.SparseTensor(idx, val, shape=[-1, self.prev_dim])
+        # import pdb; pdb.set_trace()
+        X = tf.SparseTensor(tf.cast(idx, 'int64'), val, shape=[-1, self.prev_dim])
         X = tf.sparse_reorder(X)
         XW = tf.sparse_tensor_dense_matmul(X, self.W, adjoint_a=False, adjoint_b=False)
         return tf.sparse_add(XW, self.b)
